+++ TrashMob/client-app/src/components/Partners/PartnerUsers.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react'
 import { RouteComponentProps } from 'react-router-dom';
-import UserData from './Models/UserData';
+import UserData from '../Models/UserData';
 import { Button, Col, Form, OverlayTrigger, Tooltip } from 'react-bootstrap';
-import PartnerUserData from './Models/PartnerUserData';
-import { apiConfig, getDefaultHeaders, msalClient } from '../store/AuthStore';
-import * as ToolTips from "../store/ToolTips";
+import { apiConfig, getDefaultHeaders, msalClient } from '../../store/AuthStore';
+import * as ToolTips from "../../store/ToolTips";
 
 export interface PartnerUsersDataProps extends RouteComponentProps {
     partnerId: string;
