+++ TrashMob/client-app/src/components/Partners/Partners.tsx
